requirements.txt
aniso8601==7.0.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.4.1
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.9
dj-database-url==2.0.0
Django==3.2.20
django-jet-reboot==1.3.3
django-crispy-forms==1.13.0
django-import-export==2.9.0
django-js-asset>=2.0
django-shortuuidfield==0.1.3
django-ckeditor==6.7.2
django-taggit==3.0.0
dnspython==2.6.1
email-validator==2.0.0.post2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.16.1
graphene==3.4
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
odfpy==1.4.1
openpyxl==3.1.2
Pillow==8.4.0
platformdirs==4.3.6
promise==2.3
psycopg2==2.9.9
psycopg2-binary==2.9.6
pydantic==1.10.12
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2022.7.1
pyuploadcare==4.1.0
PyYAML==6.0.1
requests==2.32.2
Rx==1.6.1
setuptools==70.2.0
shortuuid==1.0.13
singledispatch==3.6.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tablib==3.6.1
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.2.1
uvicorn==0.32.0
virtualenv==20.27.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0



